+++ components/script/dom/canvas/offscreencanvas.rs
@@ -10,7 +10,8 @@ use euclid::default::Size2D;
 use js::error::throw_type_error;
 use js::realm::CurrentRealm;
 use js::rust::{HandleObject, HandleValue};
-use pixels::{EncodedImageType, Snapshot};
+use pixels::Snapshot;
+use pixels::encoding::EncodedImageType;
 use rustc_hash::FxHashMap;
 use script_bindings::cell::{DomRefCell, Ref};
 use script_bindings::inheritance::Castable;
