+++ components/script/dom/html/embedded_content/htmlcanvaselement.rs
@@ -12,7 +12,8 @@ use js::context::NoGC;
 use js::error::throw_type_error;
 use js::rust::{HandleObject, HandleValue};
 use layout_api::HTMLCanvasData;
-use pixels::{EncodedImageType, Snapshot};
+use pixels::Snapshot;
+use pixels::encoding::EncodedImageType;
 use rustc_hash::FxHashMap;
 use script_bindings::cell::{DomRefCell, Ref};
 use script_bindings::reflector::DomObject;
